backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
sqlalchemy>=2.0
pydantic>=2.7
pydantic-settings>=2.3
email-validator>=2.2
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
python-multipart>=0.0.9
pandas>=2.2
numpy>=2.0
scikit-learn>=1.5
plotly>=5.22
chromadb>=0.5
sentence-transformers>=3.0
python-docx>=1.1
pypdf>=4.3
httpx>=0.27
groq>=0.9
